# use bitshifts for int to int in convert_dtype