@pmeier 2022-11-28
I agree this is somewhat confusing, since I needed to mix two benchmarks above. Let me try to untangle it for you:

> On your original comment (old PR), you have indicated that this approach is not faster due to the lack of vectorization on Core.

That was true at the time of writing it.

> But now you find it about 26% faster, which is not as fast as you expect. Is my understanding correct?

I find it 26% faster than v1. This is the same as for the multiplication that we currently have in v2. Meaning,

- v2 (`main`) vs. v1: +27%
- v2 (this PR) vs. v1: +26%
- v2 (this PR) vs v2 (`main`): approx 0%

> Was there work around the specific op on Core that justifies the difference between the previous attempt and now?

Yes, the ops were vectorized in core by @alexsamardzic. We are discussing offline as we speak if my benchmark in this PR results are correct or not. Will update here if we reached a conclusion.

Due to these changes on core, you see this speedup in the other branch (int to int downscale) that uses bitshifts, but wasn't touched by this PR at all.