@alexsamardzic 2022-11-28
The shift operators were not vectorized previosly.  Recently merged PRs #88607, #88990 and #89284 implement vectorization of shift operators for all integer datatypes, and this is where the performance improvement in the above benchmarks comes from.  On the other side, multiplication operator was already vectorized for some datatypes, including int32.  An operator being vectorized means that it's executed through a single (in most cases) assembly instruction, that operates on several tensor(s) elements in parallel.  Corresponding assembly instructions for shifts and multiplications are typically close in number of processor cycles required (it depends on particular processor), and results of the last benchmark are as expected in that regards.